third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/semantics/forms/the-selectlist-element/selectlist-option-focusable.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Validate <option> is focusable when is a descendant of <selectlist>
  assert_equals: expected "two" but got "one"
Harness: the test ran to completion.
